requirements.txt
beautifulsoup4
black
aiohttp
jupyter
pycodestyle
python-dotenv
selenium
tqdm
urllib3
google-api-python-client
google_auth_oauthlib